% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{export_go_results}
\alias{export_go_results}
\title{Export GO result with p-value}
\usage{
export_go_results(result_all, path, ...)
}
\arguments{
\item{result_all}{GO result}

\item{path}{path to save result}

\item{...}{
  Arguments passed on to \code{\link[=get_go_results]{get_go_results}}
  \describe{
    \item{\code{value}}{value names for export. eg. pvalue, p.adjust, qvalue, Count}
  }}
}
\description{
Export GO result with p-value
}
