% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{get_genes}
\alias{get_genes}
\title{Get GO genes}
\usage{
get_genes(go_result, all = TRUE)
}
\arguments{
\item{go_result}{GO result}

\item{all}{GO result contains all GO type or not}
}
\value{
GO result data tibble
}
\description{
Get GO genes
}
