% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{get_go_results}
\alias{get_go_results}
\title{Get GO results}
\usage{
get_go_results(result, value = "pvalue")
}
\arguments{
\item{result}{GO result}

\item{value}{value names for export. eg. pvalue, p.adjust, qvalue, Count}
}
\value{
tibble containing results
}
\description{
Get GO results
}
