% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revigo.R
\name{plot_revigo_scatter}
\alias{plot_revigo_scatter}
\title{Plot revigo results}
\usage{
plot_revigo_scatter(result, ...)
}
\arguments{
\item{result}{filtered result of revigo}

\item{...}{
  Arguments passed on to \code{\link[=plot_revigo_scatter_one]{plot_revigo_scatter_one}}
  \describe{
    \item{\code{top}}{top dispensability number for show}
  }}
}
\value{
ggplot2 object list
}
\description{
Plot revigo results
}
