% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revigo.R
\name{plot_revigo_scatter_one}
\alias{plot_revigo_scatter_one}
\title{Plot revigo results per item}
\usage{
plot_revigo_scatter_one(result_one, top = FALSE, ...)
}
\arguments{
\item{result_one}{one result item}

\item{top}{top dispensability number for show}

\item{...}{label show condition}
}
\value{
ggplot2 object
}
\description{
Plot revigo results per item
}
