% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revigo.R
\name{revigo}
\alias{revigo}
\title{Submit revigo job}
\usage{
revigo(go_result, revigo_out)
}
\arguments{
\item{go_result}{GO result with tab seperation.}

\item{revigo_out}{Output path.}
}
\description{
Submit revigo job
}
