% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tomato_all}
\alias{tomato_all}
\title{GO gson of heinz}
\format{
A gson
}
\source{
\url{https://www.sgn.cornell.edu/}
\url{http://bioinfo.cau.edu.cn/agriGO/}
}
\usage{
tomato_all
}
\description{
A gson containing the GO annotation of tomato. Gene id is 'Solyc' like.
The dataset is merged from all annotation we can find.
}
\keyword{datasets}
