% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{tomato_go}
\alias{tomato_go}
\title{Do GO analysis for tomato with all type}
\usage{
tomato_go(gene, go_type = "all", ...)
}
\arguments{
\item{go_type}{GO types}

\item{...}{
  Arguments passed on to \code{\link[clusterProfiler:enricher]{clusterProfiler::enricher}}
  \describe{
    \item{\code{gene}}{a vector of gene id}
    \item{\code{pvalueCutoff}}{adjusted pvalue cutoff on enrichment tests to report}
    \item{\code{pAdjustMethod}}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}
    \item{\code{universe}}{background genes. If missing, the all genes listed in the database (eg TERM2GENE table) will be used as background.}
    \item{\code{minGSSize}}{minimal size of genes annotated for testing}
    \item{\code{maxGSSize}}{maximal size of genes annotated for testing}
    \item{\code{qvalueCutoff}}{qvalue cutoff on enrichment tests to report as significant.  Tests must pass i) \code{pvalueCutoff} on unadjusted pvalues, ii) \code{pvalueCutoff} on adjusted pvalues and iii) \code{qvalueCutoff} on qvalues to be reported.}
  }}
}
\value{
GO retults
}
\description{
Do GO analysis for tomato with all type
}
