% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tomato_mf}
\alias{tomato_mf}
\title{GO gson of heinz}
\format{
A gson
}
\source{
\url{https://www.sgn.cornell.edu/}
\url{http://bioinfo.cau.edu.cn/agriGO/}
}
\usage{
tomato_mf
}
\description{
A gson containing the GO molecular function annotation of tomato.
Gene id is 'Solyc' like. The dataset is merged from all
annotation we can find.
}
\keyword{datasets}
